requirements.txt
asgiref==3.8.1
dj-database-url==2.3.0
Django==5.2
django-colorfield==0.14.0
django-jazzmin==3.0.1
django-tinymce==4.1.0
Markdown==3.8
pillow==11.2.1
python-decouple==3.8
sqlparse==0.5.3
typing_extensions==4.13.2
tzdata==2025.2
